Delete items in cart API

+++ src/main/java/com/betadevels/onlineshopping/payload/request/cart/DeleteItemsInCartRequest.java
@@ -16,7 +16,7 @@ public class DeleteItemsInCartRequest {
     @NotNull
     private Integer quantity;
     @NotNull
-    private int cartId;
+    private Long cartId;
 
 
 }
